+++ previous-versions/files/src/sdk/models/operations/listfiles.ts
@@ -5,10 +5,43 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import * as shared from "../shared";
 import { AxiosResponse } from "axios";
+import { Expose } from "class-transformer";
 
 export class ListFilesRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
+}
+
+/**
+ * One or more of the resources you referenced could not be found.
+ *
+ * @remarks
+ * This might be because your company or data connection id is wrong, or was already deleted.
+ */
+export class ListFilesErrorMessage extends SpeakeasyBase {
+    @SpeakeasyMetadata()
+    @Expose({ name: "canBeRetried" })
+    canBeRetried?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "correlationId" })
+    correlationId?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "detailedErrorCode" })
+    detailedErrorCode?: number;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "error" })
+    error?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "service" })
+    service?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "statusCode" })
+    statusCode?: number;
 }
 
 export class ListFilesResponse extends SpeakeasyBase {
@@ -16,10 +49,13 @@ export class ListFilesResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * One or more of the resources you referenced could not be found.
+     *
+     * @remarks
+     * This might be because your company or data connection id is wrong, or was already deleted.
      */
     @SpeakeasyMetadata()
-    errorMessage?: shared.ErrorMessage;
+    errorMessage?: ListFilesErrorMessage;
 
     /**
      * Success
@@ -32,4 +68,10 @@ export class ListFilesResponse extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    schema?: shared.Schema;
 }