+++ previous-versions/files/src/sdk/models/operations/uploadfiles.ts
@@ -5,6 +5,7 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import * as shared from "../shared";
 import { AxiosResponse } from "axios";
+import { Expose } from "class-transformer";
 
 export class UploadFilesRequestBody extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "multipart_form, content=true" })
@@ -27,19 +28,60 @@ export class UploadFilesRequest extends SpeakeasyBase {
     connectionId: string;
 }
 
+/**
+ * One or more of the resources you referenced could not be found.
+ *
+ * @remarks
+ * This might be because your company or data connection id is wrong, or was already deleted.
+ */
+export class UploadFilesErrorMessage extends SpeakeasyBase {
+    @SpeakeasyMetadata()
+    @Expose({ name: "canBeRetried" })
+    canBeRetried?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "correlationId" })
+    correlationId?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "detailedErrorCode" })
+    detailedErrorCode?: number;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "error" })
+    error?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "service" })
+    service?: string;
+
+    @SpeakeasyMetadata()
+    @Expose({ name: "statusCode" })
+    statusCode?: number;
+}
+
 export class UploadFilesResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     contentType: string;
 
     /**
-     * The request made is not valid.
+     * One or more of the resources you referenced could not be found.
+     *
+     * @remarks
+     * This might be because your company or data connection id is wrong, or was already deleted.
      */
     @SpeakeasyMetadata()
-    errorMessage?: shared.ErrorMessage;
+    errorMessage?: UploadFilesErrorMessage;
 
     @SpeakeasyMetadata()
     statusCode: number;
 
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
+
+    /**
+     * The request made is not valid.
+     */
+    @SpeakeasyMetadata()
+    schema?: shared.Schema;
 }